+++ FinBoxLending/LendingView.swift
@@ -42,8 +42,8 @@ struct LendingView_Previews: PreviewProvider {
     static var previews: some View {
         let _ = FinBoxLending.Builder()
             .apiKey(key: "f078b4d0-d171-4ed9-a1ce-c02134213b6c")
-            .customerID(id: "demo_lender_user_11221538")
-            .userToken(token: "oEuuQupgXPgGWeWSWKKBNePmXfmpcGmcIAeNCnTaYQHzOLJHFUXBPeMTHOQhOrve")
+            .customerID(id: "demo_lender_user_11281543")
+            .userToken(token: "RmvdboEyIfgpCwGlHinwWhhCTnQgQiwIkxLyXosjGZBqLOFDcDfDFcdhxTviyqOc")
             .environment(env: "UAT")
             .build()
         LendingView() {