+++ FinBoxLending/constants/Constants.swift
@@ -7,14 +7,42 @@
 
 import Foundation
 
-struct Constants {
-    // User default keys set during initialization
-    static var FINBOX_LENDING_CUSTOMER_ID = "finbox_lending_customer_id"
-    static var FINBOX_LENDING_API_KEY = "finbox_lending_api_key"
-    static var FINBOX_LENDING_USER_TOKEN = "finbox_lending_user_token"
-    static var FINBOX_LENDING_ENVIRONMENT = "finbox_lending_environment"
-    static var FINBOX_LENDING_CREDIT_LINE_AMOUNT = "finbox_lending_credit_line_amount"
-    static var FINBOX_LENDING_CREDIT_LINE_TRANSACTION_ID = "finbox_lending_credit_line_transaction_id"
-    static var FINBOX_LENDING_SHOW_TOOL_BAR = "finbox_lending_show_tool_bar"
-    static var FINBOX_LENDING_HIDE_POWERED_BY = "finbox_lending_hide_powered_by"
-}
+// User default keys set during initialization
+let FINBOX_LENDING_CUSTOMER_ID = "finbox_lending_customer_id"
+let FINBOX_LENDING_API_KEY = "finbox_lending_api_key"
+let FINBOX_LENDING_USER_TOKEN = "finbox_lending_user_token"
+let FINBOX_LENDING_ENVIRONMENT = "finbox_lending_environment"
+let FINBOX_LENDING_CREDIT_LINE_AMOUNT = "finbox_lending_credit_line_amount"
+let FINBOX_LENDING_CREDIT_LINE_TRANSACTION_ID = "finbox_lending_credit_line_transaction_id"
+let FINBOX_LENDING_SHOW_TOOL_BAR = "finbox_lending_show_tool_bar"
+let FINBOX_LENDING_HIDE_POWERED_BY = "finbox_lending_hide_powered_by"
+    
+// FinBox Lending Callback Statuses
+let FINBOX_LENDING_PERSONAL_INFO_SUBMITTED = "PERSONAL_INFO_SUBMITTED"
+let FINBOX_LENDING_EXIT = "EXIT"
+let FINBOX_LENDING_APPLICATION_COMPLETED = "APPLICATION_COMPLETED"
+let FINBOX_LENDING_PAYMENT_SUCCESSFULL = "PAYMENT_SUCCESSFULL"
+let FINBOX_LENDING_WAIT = "WAIT"
+let FINBOX_LENDING_OTP_LIMIT_EXCEEDED = "OTP_LIMIT_EXCEEDED"
+let FINBOX_LENDING_LOCATION_PERMISSION = "ESIGN_LOCATION_PERMISSION"
+let FINBOX_LENDING_SHOW_PROFILE_ICON = "SHOW_PROFILE_ICON"
+let CAMERA_PERMISSION = "CAMERA_PERMISSION"
+
+let FINBOX_RESULT_CODE = 100
+let FINBOX_JOURNEY_RESULT = "finbox_result"
+let FINBOX_RESULT_CODE_SUCCESS = "MW200"
+let FINBOX_RESULT_CODE_FAILURE = "MW500"
+let FINBOX_RESULT_CODE_ERROR = "MW400"
+let FINBOX_RESULT_CODE_CREDIT_LINE_SUCCESS = "CL200"
+let FINBOX_RESULT_CODE_CREDIT_LINE_FAILURE = "CL500"
+
+// DEV URLS
+let FINBOX_LENDING_DEV_BASE_URL = "https://lendingdev.finbox.in"
+
+// UAT URLS
+let FINBOX_LENDING_UAT_BASE_URL = "https://lendinguat.finbox.in"
+
+// PROD URLS
+let FINBOX_LENDING_PROD_BASE_URL = "https://lendingapis.finbox.in"
+
+let FINBOX_LENDING_CLIENT_SESSION_ENDPOINT = "/v1/user/clientSession"