+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 #  FinBox Lending iOS SDK
 
+## 0.0.8
+
+Enable multiple environments
+Update Cocopods Source
+
+
+## 0.0.4
+
+Make Callback Payload Codable
+
+
 ## 0.0.3
 
 BankConnect version built for IOS 13.0